fix dashboard provider icons
