+++ app/assets/javascripts/controllers/container_dashboard/util/dashboard-utils-factory.js
@@ -98,25 +98,6 @@ angular.module('miq.util').factory('dashboardUtilsFactory', function() {
     myDate = Date.parse(date);
     return isNaN(myDate) ? date : myDate
   };
-  var iconClassForProvider = function(provider) {
-    switch (provider) {
-      case "openshift":
-      case "openshift_enterprise":
-        iconClass = "pficon pficon-openshift";
-        break;
-      case "kubernetes":
-        iconClass = "pficon pficon-kubernetes";
-        break;
-      case "atomic":
-        iconClass = "product-vendor-atomic";
-        break;
-      case "atomic_enterprise":
-        iconClass = "product-vendor-atomic-enterprise";
-        break;
-    }
-
-    return iconClass;
-  };
   return {
     parseDate: parseDate,
     createProvidersStatus: createProvidersStatus,
@@ -128,7 +109,6 @@ angular.module('miq.util').factory('dashboardUtilsFactory', function() {
     createServicesStatus: createServicesStatus,
     createImagesStatus: createImagesStatus,
     createRoutesStatus: createRoutesStatus,
-    updateStatus: updateStatus,
-    iconClassForProvider: iconClassForProvider
+    updateStatus: updateStatus
   };
 });
