+++ app/assets/stylesheets/container_providers_dashboard.scss
@@ -85,7 +85,8 @@
     padding-bottom: 2px;
     width: 100%;
     text-align: center;
-    display: inline-block
+    display: inline-block;
+    height: 91px;
 }
 
 .single-provider-containers-dashboard .card-pf.card-pf-aggregate-status-mini {
